Version 1.1.0
This work adds the ability to use event-hub on both sides of a client-server interaction. Additional features include the ability to define a targetWindow to post to on emit and a fallback targetWindow if the target is never changing. New originRegex option allows users to specify a range of origins to accept messages from.